feature/sw-281-create ranking and leaderboard cards for communities
