+++ src/components/communities/CommunityDetailsClient.tsx
@@ -69,6 +69,8 @@ import PrivatePage from "../common/Overlay/PrivatePage"
 import pusherClient from "@/src/services/realtime/PusherClient"
 import { EntityUpdateBroadCast } from "@/src/utils/constants"
 import { onlineUsersStore } from "@/src/store/onlineUsers/onlineUsersStore"
+import RankingCard from "./RankingCard"
+import { CommunityRankingsData } from "../Dashboard/profile/trust-engine/Constant"
 
 interface CommunityDetailsClientProps {
   community: CommunityDetailData
@@ -114,7 +116,10 @@ export default function CommunityDetailsClient({
     "COMMUNITY",
     community?.id
   )
-
+  const user = useAtomValue(userStore.AuthUser)
+  const roleDisplay = user?.roles?.find(
+    (r) => r.role?.name?.toLowerCase() === "student"
+  )?.role?.name
   useEffect(() => {
     if (community) {
       const transformedCommunity: SelectCommunity = {
@@ -321,6 +326,8 @@ export default function CommunityDetailsClient({
   if (showAccessDeniedOverlay) {
     return <PrivatePage page="Community" pageHref="/communities" />
   }
+  const currentUserRank = CommunityRankingsData.find((r) => r.isCurrentUser)
+
   return (
     <div className="min-h-screen bg-background relative">
       {/* Added relative for the overlay positioning */}
@@ -655,6 +662,12 @@ export default function CommunityDetailsClient({
                   ))}
                 </div>
               </div>
+              {roleDisplay && (
+                <RankingCard
+                  communityTitle={community.title}
+                  currentUserRank={currentUserRank}
+                />
+              )}
             </div>
           </div>
 
@@ -747,7 +760,7 @@ export default function CommunityDetailsClient({
         open={isInviteDialogOpen}
         onOpenChange={setIsInviteDialogOpen}
         spaceName="Platform"
-        type={["link","email"]}
+        type={["link", "email"]}
         entityType="community"
         entity={community}
       />
